Added and rewrote retransmissions.md
